Progress for 6.23.2025: Notes: Fix build Errors. Today, I attempted at fixing the physics system and making it more robust, and by implemneting the ECS System. Untested: Point Light Shader, will result in error!

+++ openGLPhysics/opengl-physics/tools/include/timer.h
@@ -16,7 +16,7 @@ namespace tools
 
 		Timer(bool start);
 
-		double stop_time(bool debug = false);
+		float stop_time(bool debug = false);
 
 		void start_time();
